+++ packages/node/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Fixed
-- Adjusted filter function to stringify decoded message data of object types for correct comparison with filters. (#173)
+- Adjusted filter function to stringify decoded message data of Long types for correct comparison with filters. (#173)
 
 ## [2.10.2] - 2023-08-24
 ### Changed